BJSM continuous (snSMART with three active treatments and a continuous outcome design)

BJSM (Bayesian Joint Stage Modeling) method that borrows information across both stages to estimate the individual response rate of each treatment (with continuous outcome and a mapping function).

BJSM_c(
  data,
  xi_prior.mean,
  xi_prior.sd,
  phi3_prior.sd,
  n_MCMC_chain,
  n.adapt,
  MCMC_SAMPLE,
  ci = 0.95,
  n.digits,
  thin = 1,
  BURN.IN = 100,
  cran_check_option = FALSE,
  jags.model_options = NULL,
  coda.samples_options = NULL,
  verbose = FALSE,
  ...
)

# S3 method for BJSM_c
summary(object, ...)

# S3 method for summary.BJSM_c
print(x, ...)

# S3 method for BJSM_c
print(x, ...)

Arguments

data

trial ddatset with columns: id, trt1 (treatment 1), stage1outcome, stay (stay = 1 if patient stay on the same treatment in stage 2, otherwise stay = 0), trt2 (treatment 2), stage2outcome

xi_prior.mean

a 3-element vector of mean of the prior distributions (normal distribution) for xis (treatment effect). Please check the Details section for more explaination

xi_prior.sd

a 3-element vector of standard deviation of the prior distributions (normal distribution) for xis (treatment effect). Please check the Details section for more explaination

phi3_prior.sd

standard deviation of the prior distribution (folded normal distribution) of phi3 (if the patient stays on the same treatment, phi3 is the cumulative effect of stage 1 that occurs on the treatment longer term). Please check the Details section for more explaination

n_MCMC_chain

number of MCMC chains, default to 1

n.adapt

the number of iterations for adaptation

MCMC_SAMPLE

number of iterations for MCMC

ci

coverage probability for credible intervals, default = 0.95

n.digits

number of digits to keep in the final estimation of treatment effect

thin

thinning interval for monitors

BURN.IN

number of burn-in iterations for MCMC

cran_check_option

TRUE or FALSE. If FALSE, the algorithm will fit a model like usual. This should be the default for all model fitting. If TRUE, the model fitting is bypassed to pass CRAN check.

jags.model_options

a list of optional arguments that are passed to jags.model() function.

coda.samples_options

a list of optional arguments that are passed to coda.samples() function.

verbose

TRUE or FALSE. If FALSE, no function message and progress bar will be printed.

...

further arguments. Not currently used.

object

object to summarize.

x

object to print

Value

posterior_sample

an mcmc.list object generated through the coda.samples() function, which includes posterior samples of the link parameters and response rates generated through the MCMC process

mean_estimate

BJSM estimate of each parameter:

phi1 - lingering effect of the first treatment

phi3 - if the patient stays on the same treatment, phi3 is the cumulative effect of stage 1 that occurs on the treatment longer term

xi_j - the expected effect of treatment j, j = 1, 2, 3 in the first stage

rho is the inverse of the variance-covariance matrix of the multivariate distribution, first parameter indicates whether patient stayed on the same treatment (2) or not (1), second parameter indicates the row number of the inverse of variance-covariance matrix, and the third parameter indicates the column number of the inverse of the variance-covariance matrix

ci_estimate

x% credible interval for each parameter. By default round to 2 decimal places, if more decimals are needed, please access the results by [YourResultName]$ci_estimates$CI_low or [YourResultName]$ci_estimates$CI_high

Details

section 2.2.1 and 2.2.2 of the paper listed under reference provides a detailed description of the assumptions and prior distributions of the model.

Note that this package does not include the JAGS library, users need to install JAGS separately. Please check this page for more details: https://sourceforge.net/projects/mcmc-jags/

References

Hartman, H., Tamura, R.N., Schipper, M.J. and Kidwell, K.M., 2021. Design and analysis considerations for utilizing a mapping function in a small sample, sequential, multiple assignment, randomized trials with continuous outcomes. Statistics in Medicine, 40(2), pp.312-326.

Examples

trialData <- trialDataMF

BJSM_result <- BJSM_c(
  data = trialData, xi_prior.mean = c(50, 50, 50),
  xi_prior.sd = c(50, 50, 50), phi3_prior.sd = 20, n_MCMC_chain = 1,
  n.adapt = 1000, MCMC_SAMPLE = 5000, ci = 0.95, n.digits = 5, verbose = FALSE
)

summary(BJSM_result)
#> 
#> Parameter Estimation:
#>            Estimate   CI        CI_low      CI_high
#> phi1        0.18249 0.95  0.0006807978  0.379878116
#> phi3        3.99892 0.95  2.6841284334  5.261429201
#> rho[1,1,1]  0.00963 0.95  0.0049383247  0.014804336
#> rho[2,1,1]  0.06384 0.95  0.0332093416  0.092598324
#> rho[1,2,1] -0.00259 0.95 -0.0064856037  0.001319847
#> rho[2,2,1] -0.07026 0.95 -0.1049904289 -0.036711575
#> rho[1,1,2] -0.00259 0.95 -0.0064856037  0.001319847
#> rho[2,1,2] -0.07026 0.95 -0.1049904289 -0.036711575
#> rho[1,2,2]  0.01125 0.95  0.0056024116  0.017182083
#> rho[2,2,2]  0.09147 0.95  0.0510124565  0.135088041
#> xi_[1]     51.04569 0.95 47.1994113539 54.866772130
#> xi_[2]     62.05141 0.95 58.0385859130 65.652831191
#> xi_[3]     69.02606 0.95 65.2219700790 72.586008100
print(BJSM_result)
#> 
#> Treatment Effect Estimation:
#>      Estimate   CI   CI_low  CI_high
#> trtA 51.04569 0.95 47.19941 54.86677
#> trtB 62.05141 0.95 58.03859 65.65283
#> trtC 69.02606 0.95 65.22197 72.58601