LPJSM for snSMART with binary outcomes (3 active treatments or placebo and two dose level)

A joint-stage regression model (LPJSM) is a frequentist modeling approach that incorporates the responses of both stages as repeated measurements for each subject. Generalized estimating equations (GEE) are used to estimate the response rates of each treatment. The marginal response rates for each DTR can also be obtained based on the GEE results.

LPJSM_binary(data, six = TRUE, DTR = TRUE, cran_check_option = FALSE, ...)

# S3 method for LPJSM_binary
summary(object, ...)

# S3 method for summary.LPJSM_binary
print(x, ...)

# S3 method for LPJSM_binary
print(x, ...)

Arguments

data

dataset with columns named as treatment_stageI, response_stageI, treatment_stageII and response_stageII

six

if TRUE, will run the six beta model, if FALSE will run the two beta model. Default is six = TRUE

DTR

if TRUE, will also return the expected response rate and its standard error of dynamic treatment regimens

cran_check_option

TRUE or FALSE. If FALSE, the algorithm will fit a model like usual. This should be the default for all model fitting. If TRUE, the model fitting is bypassed to pass CRAN check.

...

further arguments. Not currently used.

object

object to print

x

object to summarize.

Value

a list containing

GEE_output

- original output of the GEE (geeglm) model

pi_hat

- estimate of response rate/treatment effect

sd_pi_hat

- standard error of the response rate

pi_DTR_hat

- expected response rate of dynamic treatment regimens (DTRs)

pi_DTR_se

- standard deviation of DTR estimates

References

Wei, B., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2018. A Bayesian analysis of small n sequential multiple assignment randomized trials (snSMARTs). Statistics in medicine, 37(26), pp.3723-3732.

Chao, Y.C., Trachtman, H., Gipson, D.S., Spino, C., Braun, T.M. and Kidwell, K.M., 2020. Dynamic treatment regimens in small n, sequential, multiple assignment, randomized trials: An application in focal segmental glomerulosclerosis. Contemporary clinical trials, 92, p.105989.

Fang, F., Hochstedler, K.A., Tamura, R.N., Braun, T.M. and Kidwell, K.M., 2021. Bayesian methods to compare dose levels with placebo in a small n, sequential, multiple assignment, randomized trial. Statistics in Medicine, 40(4), pp.963-977.

See also

Examples

data <- data_binary

LPJSM_result <- LPJSM_binary(data = data, six = TRUE, DTR = TRUE)

summary(LPJSM_result)
#> 
#> GEE output:
#> 
#> Call:
#> geepack::geeglm(formula = Y ~ alphaA + alphaB + alphaC + gamma1A + 
#>     gamma2A + gamma1B + gamma2B + gamma1C + gamma2C - 1, family = poisson(link = "log"), 
#>     data = geedata, id = ptid, corstr = "independence")
#> 
#>  Coefficients:
#>         Estimate  Std.err     Wald Pr(>|W|)    
#> alphaA   -1.2155   0.4030    9.096  0.00256 ** 
#> alphaB   -0.9845   0.3403    8.370  0.00381 ** 
#> alphaC   -0.8444   0.3036    7.737  0.00541 ** 
#> gamma1A   1.2155   0.4030    9.096  0.00256 ** 
#> gamma2A   0.7029   0.3664    3.680  0.05508 .  
#> gamma1B   0.9845   0.3403    8.370  0.00381 ** 
#> gamma2B   0.8271   0.3497    5.594  0.01803 *  
#> gamma1C   0.8444   0.3036    7.737  0.00541 ** 
#> gamma2C -42.2449   0.6085 4820.409  < 2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Correlation structure = independence 
#> Estimated Scale Parameters:
#> 
#>             Estimate Std.err
#> (Intercept)   0.4096  0.1259
#> Number of clusters:   30  Maximum cluster size: 2 
#> 
#> Treatment Effect Estimate:
#>      Estimate Std. Error
#> trtA   0.2966     0.1195
#> trtB   0.3736     0.1271
#> trtC   0.4298     0.1305
#> 
#> Expected Response Rate of Dynamic Treatment Regimens (DTR):
#>        Estimate Std. Error
#> rep_AB   0.8274     0.1328
#> rep_AC   0.9072     0.2341
#> rep_BA   0.7984     0.1403
#> rep_BC   0.9892     0.1427
#> rep_CA   0.4298     0.1305
#> rep_CB   0.4298     0.1305
#>