BJSM method for interim analysis and final analysis of group sequential trial design

After obtain real trial data, this function can be used to decide which arm to drop in an interim analysis or provide a full final analysis.

group_seq(
  data,
  interim = TRUE,
  drop_threshold_pair = NULL,
  prior_dist,
  pi_prior,
  beta_prior,
  MCMC_SAMPLE,
  n.adapt,
  thin = 1,
  BURN.IN = 100,
  n_MCMC_chain,
  ci = 0.95,
  DTR = TRUE,
  cran_check_option = FALSE,
  jags.model_options = NULL,
  coda.samples_options = NULL,
  verbose = FALSE,
  ...
)

# S3 method for summary.group_seq
print(x, ...)

# S3 method for group_seq
print(x, ...)

Arguments

data

dataset should include 8 columns: time.1st.trt (first treatment starts time), time.1st.resp (first response time), time.2nd.trt (second treatment starts time), time.2nd.resp (second response time), trt.1st (treatment arm for first treatment), resp.1st (response for first treatment), trt.2nd (treatment arm for second treatment), resp.2nd (response for second treatment) data yet to be observed should be marked as "NA"

interim

indicates whether user is conducting an interim analysis via BJSM (interim = TRUE) or an final analysis via BJSM (interim = FALSE)

drop_threshold_pair

a vector of 2 values (drop_threshold_tau_l, drop_threshold_psi_l). Both drop_threshold_tau_l and drop_threshold_psi_l should be between 0 and 1. only assign value to this parameter when interim = TRUE. See the details section for more explanation

prior_dist

vector of three values ("prior distribution for pi", "prior distribution for beta0", "prior distribution for beta1"), user can choose from "gamma", "beta", "pareto". e.g. prior_dist = c("beta", "beta", "pareto")

pi_prior

vector of six values (a, b, c, d, e, f), where a and b are the parameter a and parameter b of the prior distribution for pi_1A, c and d are the parameter a and parameter b of the prior distribution for pi_1B, and e and f are the parameter a and parameter b of the prior distribution for pi_1C. Please check the Details section for more explanation

beta_prior

vector of four values (beta0_prior.a, beta0_prior.b, beta1_prior.a, beta1_prior.c). beta0_prior.a is the parameter a of the prior distribution for linkage parameter beta0. beta0_prior.b is the parameter b of the prior distribution for linkage parameter beta0. beta1_prior.a is the parameter a of the prior distribution for linkage parameter beta1. beta1_prior.c is the parameter b of the prior distribution for linkage parameter beta1. Please check the Details section for more explanation

MCMC_SAMPLE

number of iterations for MCMC

n.adapt

the number of iterations for adaptation

thin

thinning interval for monitors

BURN.IN

number of burn-in iterations for MCMC

n_MCMC_chain

number of MCMC chains, default to 1

ci

coverage probability for credible intervals, default = 0.95. only assign value to this parameter when interim = FALSE.

DTR,

if TRUE, will also return the expected response rate of dynamic treatment regimens. default = TRUE. only assign value to this parameter when interim = FALSE.

cran_check_option

TRUE or FALSE. If FALSE, the algorithm will fit a model like usual. This should be the default for all model fitting. If TRUE, the model fitting is bypassed to pass CRAN check.

jags.model_options

a list of optional arguments that are passed to jags.model() function.

coda.samples_options

a list of optional arguments that are passed to coda.samples() function.

verbose

TRUE or FALSE. If FALSE, no function message and progress bar will be printed.

...

further arguments. Not currently used.

x

object to summarize.

Value

if interim = TRUE, this function returns either 0 - no arm is dropped, or A/B/C - arm A/B/C is dropped

if interim = FALSE, this function returns:

posterior_sample

an mcmc.list object generated through the coda.samples() function, which includes posterior samples of the link parameters and response rates generated through the MCMC process

pi_hat_bjsm

estimate of response rate/treatment effect

se_hat_bjsm

standard error of the response rate

ci_pi_A, ci_pi_B, ci_pi_C

x% credible intervals for treatment A, B, C

diff_AB, diff_BC. diff_AC

estimate of differences between treatments A and B, B and C, A and C

ci_diff_AB, ci_diff_BC, ci_diff_AC

x% credible intervals for the differences between treatments A and B, B and C, A and C

se_AB, se_BC, se_AC

standard error for the differences between treatments A and B, B and C, A and C

beta0_hat, beta1_hat

linkage parameter beta0 and beta1 estimates

se_beta0_hat, se_beta1_hat

standard error of the estimated value of linkage parameter beta0 and beta1

ci_beta0_hat, ci_beta1_hat

linkage parameter beta0 and beta1 credible interval

pi_DTR_est

expected response rate of dynamic treatment regimens (DTRs)

pi_DTR_se

standard error for the estimated DTR response rate

ci_pi_AB, ci_pi_AC, ci_pi_BA, ci_pi_BC, ci_pi_CA, ci_pi_CB

x% credible intervals for the estimated DTR response rate

Details

For gamma distribution, prior.a is the shape parameter r, prior.b is the rate parameter lambda. For beta distribution, prior.a is the shape parameter a, prior.b is the shape parameter b. For pareto distribution, prior.a is the scale parameter alpha, prior.b is the shape parameter c (see page 29 of the jags user manual version 3.4.0). link: http://www.stats.ox.ac.uk/~nicholls/MScMCMC14/jags_user_manual.pdf

The individual response rate is regarded as a permanent feature of the treatment. The second stage outcome is modeled conditionally on the first stage results linking the first and second stage response probabilities through linkage parameters.

(paper provided in the reference section, section 2.2.2 Bayesian decision rules. drop_threshold_tau_l and drop_threshold_psi_l correspond to \(tau_l\) and \(psi_l\) respectively)

Please refer to the paper listed under reference section for detailed definition of parameters. Note that this package does not include the JAGS library, users need to install JAGS separately. Please check this page for more details: https://sourceforge.net/projects/mcmc-jags/

References

Chao, Y.C., Braun, T.M., Tamura, R.N. and Kidwell, K.M., 2020. A Bayesian group sequential small n sequential multiple‐assignment randomized trial. Journal of the Royal Statistical Society: Series C (Applied Statistics), 69(3), pp.663-680.

Examples

mydata <- groupseqDATA_look1

result1 <- group_seq(
  data = mydata, interim = TRUE, drop_threshold_pair = c(0.5, 0.4),
  prior_dist = c("beta", "beta", "pareto"), pi_prior = c(0.4, 1.6, 0.4, 1.6, 0.4, 1.6),
  beta_prior = c(1.6, 0.4, 3, 1), MCMC_SAMPLE = 6000, n.adapt = 1000, n_MCMC_chain = 1
)
#> 
#> Interim Analysis Outcome:
#> Threshold tau_l is set to: 
#> 0.5
#> 
#> Threshold psi_l is set to: 
#> 0.4
#> 
#> Step 1: Arm C's interim posterior probability of having the greatest response is bigger than threshold 
#> 0.5
#> 
#> Step 2: Arm A's interim posterior probability of having the lowest response is higher
#> Arm A is dropped
#> 

summary(result1)
#> Arm A is dropped
#> 
#> Iterations = 1101:7100
#> Thinning interval = 1 
#> Number of chains = 1 
#> Sample size per chain = 6000 
#> 
#> 1. Empirical mean and standard deviation for each variable,
#>    plus standard error of the mean:
#> 
#>          Mean     SD Naive SE Time-series SE
#> beta[1] 0.784 0.2111  0.00273        0.00721
#> beta[2] 1.245 0.2630  0.00339        0.00887
#> beta[3] 0.695 0.2812  0.00363        0.00810
#> beta[4] 1.334 0.2968  0.00383        0.00781
#> beta[5] 0.839 0.1922  0.00248        0.00784
#> beta[6] 1.221 0.2106  0.00272        0.00592
#> pi[1]   0.290 0.0883  0.00114        0.00162
#> pi[2]   0.423 0.0822  0.00106        0.00175
#> pi[3]   0.564 0.1112  0.00143        0.00233
#> 
#> 2. Quantiles for each variable:
#> 
#>          2.5%   25%   50%   75% 97.5%
#> beta[1] 0.294 0.643 0.849 0.968 1.000
#> beta[2] 1.006 1.067 1.163 1.333 1.939
#> beta[3] 0.103 0.483 0.770 0.955 1.000
#> beta[4] 1.009 1.110 1.252 1.473 2.089
#> beta[5] 0.321 0.748 0.921 0.986 1.000
#> beta[6] 1.006 1.066 1.159 1.312 1.780
#> pi[1]   0.136 0.226 0.283 0.347 0.478
#> pi[2]   0.269 0.365 0.420 0.476 0.592
#> pi[3]   0.343 0.489 0.566 0.643 0.774
#> 

mydata <- groupseqDATA_full
result2 <- group_seq(
  data = mydata, interim = FALSE, prior_dist = c("beta", "beta", "pareto"),
  pi_prior = c(0.4, 1.6, 0.4, 1.6, 0.4, 1.6),
  beta_prior = c(1.6, 0.4, 3, 1), MCMC_SAMPLE = 6000, n.adapt = 1000,
  n_MCMC_chain = 1, ci = 0.95, DTR = TRUE
)

summary(result2)
#> 
#> Treatment Effects Estimate:
#>      Estimate Std. Error C.I. CI low CI high
#> trtA   0.3014    0.04843 0.95 0.2118  0.3989
#> trtB   0.4725    0.04050 0.95 0.3934  0.5498
#> trtC   0.6469    0.08166 0.95 0.4764  0.7692
#> 
#> Differences between Treatments:
#>        Estimate Std.Error C.I. CI low   CI high
#> diffAB  -0.1711   0.06272 0.95 -0.297 -0.054807
#> diffBC  -0.1745   0.08996 0.95 -0.342  0.007595
#> diffAC  -0.3455   0.09206 0.95 -0.509 -0.153378
#> 
#> Linkage Parameter Estimate:
#>        Estimate Std. Error C.I. CI low CI high
#> beta0A   0.8794     0.1087 0.95 0.6705   1.000
#> beta0B   0.7565     0.1305 0.95 0.5555   1.000
#> beta0C   0.8825     0.1257 0.95 0.6185   1.000
#> beta1A   1.4657     0.3547 0.95 1.0002   2.143
#> beta1B   1.3641     0.1761 0.95 1.0346   1.702
#> beta1C   1.5550     0.2475 0.95 1.2706   2.042
#>